Add BaseBigQueryConnectorSmokeTest
